chore: Prettier code formatting
